requirements.txt
blinker==1.7.0
Bootstrap_Flask==2.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pip==23.2.1
python-dotenv==1.0.1
setuptools==65.5.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.1
WTForms==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.6
requests==2.26.0
